refactor: consistent naming
