+++ packages/react-ui/components/Gapped/HorizontalGapped/HorizontalGapped.tsx
@@ -3,19 +3,19 @@ import React, { forwardRef } from 'react';
 import { GappedProps, GappedRef } from '../Gapped';
 import { GapValue } from '../useGapValue';
 
-import { HorizontalItems } from './HorizontalItems';
-import { HorizontalWrapper } from './HorizontalWrapper';
+import { HorizontalGappedItems } from './HorizontalGappedItems';
+import { HorizontalGappedWrapper } from './HorizontalGappedWrapper';
 
 export type HorizontalGappedProps = Omit<GappedProps, 'vertical' | 'gap'> & GapValue;
 
 export const HorizontalGapped = forwardRef<GappedRef['element'], HorizontalGappedProps>(
   ({ gap, wrap, children, verticalAlign, ...rest }, ref) => {
     return (
-      <HorizontalWrapper ref={ref} gap={gap} wrap={wrap} {...rest}>
-        <HorizontalItems wrap={wrap} verticalAlign={verticalAlign} gap={gap}>
+      <HorizontalGappedWrapper ref={ref} gap={gap} wrap={wrap} {...rest}>
+        <HorizontalGappedItems wrap={wrap} verticalAlign={verticalAlign} gap={gap}>
           {children}
-        </HorizontalItems>
-      </HorizontalWrapper>
+        </HorizontalGappedItems>
+      </HorizontalGappedWrapper>
     );
   },
 );
