+++ packages/react-ui/components/Gapped/HorizontalGapped/HorizontalGappedItems.tsx
@@ -4,18 +4,18 @@ import { filterChildren } from '../filterChildren';
 import { GappedProps } from '../Gapped';
 import { GapValue } from '../useGapValue';
 
-import { HorizontalItem } from './HorizontalItem';
+import { HorizontalGappedItem } from './HorizontalGappedItem';
 
 export type HorizontalItemsProps = Pick<GappedProps, 'children' | 'wrap' | 'verticalAlign'> & GapValue;
 
-export const HorizontalItems = ({ gap, children, wrap, verticalAlign }: HorizontalItemsProps) => {
+export const HorizontalGappedItems = ({ gap, children, wrap, verticalAlign }: HorizontalItemsProps) => {
   return React.Children.toArray(children)
     .filter(filterChildren)
     .map((child, index) => {
       return (
-        <HorizontalItem key={index} verticalAlign={verticalAlign} gap={gap} wrap={wrap} isFirst={index === 0}>
+        <HorizontalGappedItem key={index} verticalAlign={verticalAlign} gap={gap} wrap={wrap} isFirst={index === 0}>
           {child}
-        </HorizontalItem>
+        </HorizontalGappedItem>
       );
     }) as unknown as JSX.Element;
 };
